+++ api/src/main/java/org/openmrs/module/lamp/ProgramStrategy.java
@@ -1,0 +1,10 @@
+package org.openmrs.module.lamp;
+
+import java.util.Date;
+import org.openmrs.Encounter;
+import org.openmrs.User;
+
+public interface ProgramStrategy {
+	
+	void execute(Encounter encounter, User currentUser, Date currentDate, String reason);
+}
